+++ Tests/Recurly/Coupon_Redemption_Test.php
@@ -18,19 +18,11 @@ class Recurly_CouponRedemptionTest extends Recurly_TestCase
     $this->assertInstanceOf('Recurly_Stub', $redemption->account);
     $this->assertEquals('https://api.recurly.com/v2/coupons/special', $redemption->coupon->getHref());
     $this->assertEquals('https://api.recurly.com/v2/accounts/abcdef1234567890', $redemption->account->getHref());
+    $this->assertEquals('https://api.recurly.com/v2/subscriptions/012345678901234567890123456789ab', $redemption->subscription->getHref());
     $this->assertFalse($redemption->single_use);
     $this->assertEquals(0, $redemption->total_discounted_in_cents);
     $this->assertEquals('USD', $redemption->currency);
     $this->assertEquals('active', $redemption->state);
     $this->assertEquals(1435591848, $redemption->created_at->getTimestamp());
-  }
-
-  public function testDeleteRedemption() {
-    $this->client->addResponse(
-      'DELETE', 'https://api.recurly.com/v2/accounts/abcdef1234567890/redemption',
-      'accounts/redemption/destroy-204.xml'
-    );
-
-    Recurly_CouponRedemption::get('abcdef1234567890', $this->client)->delete();
   }
 }